fix:same precision on coverage rates of current and diff branch (#49)
Minor fix following
https://github.com/insightsengineering/coverage-action/pull/47.
Addresses the following:

- Coverage rate being saved for the current branch wasn't following the
`1e-8` precision used in the diff branch;
- Removed unneeded `tr -d '%'`. 

My bad on the suggestion I left on the PR missing these points 🥲

As an additional note, I thought about adding tests for these edge cases
(80% on baseline, 80.000001% / 79.999999% on branches). However, it
seems that `pycobertura show` counts line for each `<line />` element in
the xml, so the mock files would require 1e8 lines to do so, which is
far from ideal. Due to that, tests were left untouched.